New Features
Now  the device will serial print the MAC address of its wifi device.  Also it now has configurable timouts on connection attempts for wifi and mqtt server connections in the event that it looses connection to any of these servers it will make the configured attempts and then go to low power mode.